requirements.txt
bcrypt==4.3.0
ecdsa==0.19.1
pyasn1==0.4.8
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9.1
six==1.17.0
